requirements.txt
Python >= 3.7
pytorch==1.8.0 
torchvision==0.9.0 
torchaudio==0.8.0 
cudatoolkit=11.1

